requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
diffusers==0.6.0
python-dotenv==0.21.0
Django==4.1.2
django-binary-database-files==1.0.17
django-cors-headers==3.13.0
django-shortcuts==1.6
executing==1.2.0
filelock==3.8.0
ftfy==6.1.1
grpcio==1.48.1
grpcio-tools==1.48.1
gunicorn==20.1.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.23.4
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3
postgres==4.0
prompt-toolkit==3.0.32
protobuf==3.19.5
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
stability-sdk==0.3.1
stack-data==0.6.0
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
zipp==3.9.0
